+++ ionWindow/CEventManager.cpp
@@ -61,9 +61,9 @@ void CApplicationEventReceiver::OnApplicationExit()
 CEventManager::CEventManager()
     : IsKeyDown(KeyStates), MouseLocation(MousePositionState), IsMouseDown(MouseStates)
 {
-	for (unsigned int i = 0; i < EKey::Count; ++ i)
+	for (unsigned int i = 0; i < (int) EKey::Count; ++ i)
         KeyStates[i] = false;
 
-    for (unsigned int i = 0; i < SMouseEvent::EButton::Count; ++ i)
+    for (unsigned int i = 0; i < (int) SMouseEvent::EButton::Count; ++ i)
         MouseStates[i] = false;
 }
